requirements.txt
numpy
pandas
ipython
matplotlib
seaborn
tqdm
opencv-python
albumentations
timm
efficientnet-pytorch